+++ package.json
@@ -22,12 +22,12 @@
 		"@fontsource/roboto": "^5.2.9",
 		"@fontsource/spectral": "^5.2.8",
 		"@lucide/svelte": "^0.555.0",
-		"@maiertech/sveltekit-helpers": "^0.40.1",
+		"@maiertech/sveltekit-helpers": "^0.41.2",
 		"@playwright/test": "^1.57.0",
 		"@stackblitz/sdk": "^1.11.0",
 		"@sveltejs/adapter-node": "^5.4.0",
 		"@sveltejs/enhanced-img": "^0.9.2",
-		"@sveltejs/kit": "^2.49.0",
+		"@sveltejs/kit": "^2.49.1",
 		"@sveltejs/vite-plugin-svelte": "^6.2.1",
 		"@tailwindcss/vite": "^4.1.17",
 		"@types/node": "^24.10.1",
@@ -39,19 +39,19 @@
 		"globals": "^16.5.0",
 		"mdsvex": "^0.12.6",
 		"mode-watcher": "^1.1.0",
-		"prettier": "3.6.2",
+		"prettier": "^3.7.4",
 		"prettier-plugin-svelte": "^3.4.0",
-		"prettier-plugin-tailwindcss": "^0.7.1",
-		"shiki": "^3.17.0",
-		"simple-icons": "^15.22.0",
-		"svelte": "^5.45.2",
+		"prettier-plugin-tailwindcss": "^0.7.2",
+		"shiki": "^3.19.0",
+		"simple-icons": "^16.0.0",
+		"svelte": "^5.45.5",
 		"svelte-check": "^4.3.4",
 		"sveltekit-embed": "^0.0.22",
 		"tailwind-merge": "^3.4.0",
 		"tailwindcss": "^4.1.17",
 		"typescript": "^5.9.3",
-		"typescript-eslint": "^8.48.0",
-		"vite": "^7.2.4",
+		"typescript-eslint": "^8.48.1",
+		"vite": "^7.2.6",
 		"vite-plugin-devtools-json": "^1.0.0",
 		"zod": "^4.1.13"
 	},
