+++ src/lib/server/collections/tags.ts
@@ -1,5 +1,5 @@
 import { allTags } from 'content-collections';
 
-export const all = allTags.toSorted((a, b) => {
+export const sorted = allTags.toSorted((a, b) => {
 	return a.label.localeCompare(b.label);
 });
