+++ src/routes/posts/+page.server.ts
@@ -1,4 +1,4 @@
-import { all as posts } from '$lib/server/collections/posts';
+import { sorted as posts } from '$lib/server/collections/posts.js';
 import type { PageServerLoad } from './$types';
 
 export const prerender = true;
