+++ src/routes/videos/(2025)/vs-code-finally-supports-git-worktrees/+page.md
@@ -1,0 +1,36 @@
+---
+id: x1mjUh6-azc
+---
+
+Ever been stuck switching between branches when you need to work on multiple issues? Git worktrees
+solves this problem perfectly!
+
+## What You'll Learn:
+
+✅ How to create and manage Git worktrees in VS Code ✅ Work on multiple branches without stashing
+or committing unfinished work ✅ Set up separate workspaces for different features/issues ✅ Create
+pull requests and merge workflows with worktrees ✅ Clean up and delete worktrees when done
+
+## Key Benefits:
+
+🚀 _Boost Productivity_ - No more stashing changes or switching branches ⚡ _Parallel Development_ -
+Work on multiple issues simultaneously 🎯 _Clean Workflow_ - Each worktree is an independent
+workspace 🔄 _Easy Switching_ - Jump between different features effortlessly
+
+## Timestamps:
+
+00:00 What are Git Worktrees? 01:11 Limitations of native Git worktree support in VS Code v1.103
+01:32 Git Worktree Menu extension 02:01 From GitHub issue to Git worktree 03:03 What Git worktrees
+look like on the file system 04:20 Open a Git worktree in VS Code 04:31 Git worktrees are separate
+workspaces 05:43 Creating a pull request 06:24 Deleting a Git worktree 08:04 The
+`git.detectWorktrees` setting 09:08 Don't forget to pull you changes into `main` 09:35 Wrap up
+
+## Links
+
+- VS Code v1.103 release notes: https://code.visualstudio.com/updates/v1_103#_git-worktree-support
+- VS Code worktrees docs: https://code.visualstudio.com/docs/sourcecontrol/overview#_worktrees
+- Git Worktree Menu extension:
+  https://marketplace.visualstudio.com/items?itemName=CodeInKlingon.git-worktree-menu
+
+#VSCode #Git #GitWorktrees #DeveloperProductivity #Coding #VisualStudioCode #GitTips #WebDevelopment
+#SoftwareDevelopment
